Initial notebook for k-means clustering algorithm
The algorithm has been implemented, but it would be good to calculate
the variance of each cluster so that we can look at deciding a good
value for `k` based on the variance delta leveling off.

Also, it would be good to visualise the results of the algorithm to
better understand the output.